Added live chat functionality
Added the ability to live chat with users.
If the recipient is not online, the socket is not used, but the message will still be sent to the database.

if the user is connected, all information is sent through a socket, including whether or not the user is typing.

Additional work required, but base functionality is there.