Added admin functions
Added Admin requirements to functions such as Admin emit to alert users a volunteer wishes to talk and the ability to view the panel.
Also added a session check to prevent socket trying to establish a connection before this was created.
Added a new section in Users database that specifies the last time they were contacted by a Volunteer. When connection is made with a volunteer, this is updated.
All users in the Admin panel will appear in Ascending order from oldest to newest, in order to maintain a fair rotation.