Various updates
Improved responsiveness for mobile and desktop.
Changed profile layout to be be more appealing and in line with overall aesthetics.
Removed upload profile picture from profile page to edit profile.
Added Logo to Navbar.
Added dynamic div sizing based on selected font size preference.
Changed various elements of the application to improve aesthetics.
When a user is sent a chat request from an admin, the admin is alerted if the user has joined or declined the chat request.